backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
cryptography==43.0.1
apscheduler==3.10.4
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
hypothesis==6.112.2
email-validator==2.2.0
python-multipart==0.0.12
